requirement.txt
flask-cors==6.0.2
Flask==3.1.3
Flask-Compress==1.23
dotenv==0.9.9
databricks-sdk==0.73.0
databricks-sdk[openai]
databricks-sql-connector==4.1.3